+++ inlang/source-code/paraglide/paraglide-js-adapter-sveltekit/example/src/lib/ui/Header.svelte
@@ -6,7 +6,7 @@
 <header>
     <nav>
         <a href="/" aria-current={$page.route.id === "/" ? "page" : undefined}>{m.home()}</a>
-        <a href="/about" aria-current={$page.route.id === "/about" ? "page" : undefined}>{m.about()}</a>
+        <a href="/about/" aria-current={$page.route.id === "/about" ? "page" : undefined}>{m.about()}</a>
         <a href="/users/1" aria-current={$page.route.id === "/users/[id]" ? "page" : undefined}>{m.users()}</a>
     </nav>
     <span>{$page.url.pathname}</span>