+++ inlang/source-code/paraglide/paraglide-js-sveltekit-old/src/adapter-sveltekit/examples/lang-in-slug/src/routes/[lang]/+page.server.ts
@@ -1,6 +1,0 @@
-import { initServerLoadWrapper } from "@inlang/paraglide-js-sveltekit/adapter-sveltekit/server"
-import type { PageServerLoad } from "./$types.js"
-
-export const load = initServerLoadWrapper<PageServerLoad>().use(async (_, { i }) => {
-	console.info("[lang]/+page.server.ts", i("welcome"))
-})