+++ inlang/source-code/paraglide/paraglide-js-sveltekit-old/src/adapter-sveltekit/examples/spa/src/routes/about/+layout.svelte
@@ -1,9 +1,0 @@
-<script lang="ts">
-	import { getRuntimeFromContext } from "@inlang/paraglide-js-sveltekit/adapter-sveltekit/client/reactive"
-
-	const { i } = getRuntimeFromContext()
-
-	console.info("about/+layout.svelte", $i("welcome"))
-</script>
-
-<slot />