+++ inlang/source-code/paraglide/paraglide-js-sveltekit-old/src/detectors/detectors/shared/index.ts
@@ -1,1 +1,0 @@
-export { initRootSlugDetector } from "./rootSlugDetector.js"